# open-api/immich-openapi-specs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/libraries": {
      "get": {
        "operationId": "getAllLibraries",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LibraryResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Libraries"
        ]
      },
      "post": {
        "operationId": "createLibrary",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLibraryDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Libraries"
        ]
      }
    },
    "/libraries/{id}": {
      "delete": {
        "operationId": "deleteLibrary",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Libraries"
        ]
      },
      "get": {
        "operationId": "getLibrary",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Libraries"
        ]
      },
      "put": {
        "operationId": "updateLibrary",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLibraryDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Libraries"
        ]
      }
    },
    "/libraries/{id}/scan": {
      "post": {
        "operationId": "scan",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanLibraryDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Libraries"
        ]
      }
    },
    "/libraries/{id}/statistics": {
      "get": {
        "operationId": "getLibraryStatistics",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryStatsResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Libraries"
        ]
      }
    },
    "/libraries/{id}/validate": {
      "post": {
        "operationId": "validate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateLibraryDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateLibraryResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Libraries"
        ]
      }
    },
...
  },
  "info": {
    "title": "Immich",
    "description": "Immich API",
    "version": "1.115.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "Bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "immich_access_token"
      },
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "CreateLibraryDto": {
        "properties": {
          "exclusionPatterns": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "importPaths": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "name": {
            "type": "string"
          },
          "ownerId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "ownerId"
        ],
        "type": "object"
      },
...
      "LibraryResponseDto": {
        "properties": {
          "assetCount": {
            "type": "integer"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "exclusionPatterns": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "id": {
            "type": "string"
          },
          "importPaths": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "name": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "refreshedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "assetCount",
          "createdAt",
          "exclusionPatterns",
          "id",
          "importPaths",
          "name",
          "ownerId",
          "refreshedAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "LibraryStatsResponseDto": {
        "properties": {
          "photos": {
            "default": 0,
            "type": "integer"
          },
          "total": {
            "default": 0,
            "type": "integer"
          },
          "usage": {
            "default": 0,
            "format": "int64",
            "type": "integer"
          },
          "videos": {
            "default": 0,
            "type": "integer"
          }
        },
        "required": [
          "photos",
          "total",
          "usage",
          "videos"
        ],
        "type": "object"
      },
...
      "ScanLibraryDto": {
        "properties": {
          "refreshAllFiles": {
            "type": "boolean"
          },
          "refreshModifiedFiles": {
            "type": "boolean"
          },
          "removeDeleted": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "UpdateLibraryDto": {
        "properties": {
          "exclusionPatterns": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "importPaths": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "name": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ValidateLibraryDto": {
        "properties": {
          "exclusionPatterns": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "importPaths": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "ValidateLibraryImportPathResponseDto": {
        "properties": {
          "importPath": {
            "type": "string"
          },
          "isValid": {
            "default": false,
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "importPath",
          "isValid"
        ],
        "type": "object"
      },
      "ValidateLibraryResponseDto": {
        "properties": {
          "importPaths": {
            "items": {
              "$ref": "#/components/schemas/ValidateLibraryImportPathResponseDto"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
    }
  }
}